feat: add react package
