+++ package.json
@@ -49,7 +49,9 @@
     "resolutions": {
         "@atview/docs>@atview/core": "workspace:*",
         "@atview/docs>@atview/ui": "workspace:*",
-        "@atview/docs>@atview/xrpc": "workspace:*"
+        "@atview/docs>@atview/xrpc": "workspace:*",
+        "@atview/docs>@atview/react": "workspace:*",
+        "@atview/react>@atview/core": "workspace:*"
     },
     "license": "MIT",
     "packageManager": "pnpm@11.0.0-dev.1005+sha512.91f84a392eea348ea4852a182912d2520273a4336f933b78cc44bc931eb999923c097e9433a9b355adc1f725725ea99082fc9f032a559df832632e764c92c798"
