+++ packages/react/package.json
@@ -1,7 +1,7 @@
 {
-  "name": "@atview/docs",
+  "name": "@atview/react",
   "version": "0.0.1",
-  "description": "Documents features for atproto text-based content",
+  "description": "React tools for atproto text-based content",
   "files": [
     "dist"
   ],
@@ -12,14 +12,11 @@
       "require": "./dist/cjs/index.js",
       "default": "./dist/esm/index.mjs"
     },
-    "./client": {
-      "types": "./dist/cjs/index.client.d.ts",
-      "import": "./dist/esm/index.client.mjs",
-      "require": "./dist/cjs/index.client.js",
-      "default": "./dist/esm/index.client.mjs"
-    },
-    "./styles.css": {
-      "default": "./dist/styles.css"
+    "./default": {
+      "types": "./dist/cjs/default.d.ts",
+      "import": "./dist/esm/default.mjs",
+      "require": "./dist/cjs/default.js",
+      "default": "./dist/esm/default.mjs"
     }
   },
   "scripts": {
@@ -28,9 +25,8 @@
   },
   "keywords": [
     "atproto",
-    "documents",
-    "components",
-    "react"
+    "react",
+    "tools"
   ],
   "repository": {
     "type": "git",
@@ -42,40 +38,22 @@
     "url": "https://github.com/alexdln/"
   },
   "license": "MIT",
-  "dependencies": {
-    "@atproto/api": "0.19.3",
-    "clsx": "2.1.1",
-    "contection": "2.3.1",
-    "contection-top-layer": "2.3.1",
-    "hast-util-to-jsx-runtime": "2.3.6",
-    "katex": "0.16.45",
-    "shiki": "4.0.2"
-  },
   "devDependencies": {
     "@atview/core": "./atview-core-0.0.1.tgz",
-    "@atview/ui": "./atview-ui-0.0.1.tgz",
-    "@atview/react": "./atview-react-0.0.1.tgz",
-    "@atview/xrpc": "./atview-xrpc-0.0.1.tgz",
     "@rollup/plugin-commonjs": "29.0.2",
     "@rollup/plugin-node-resolve": "16.0.3",
     "@rollup/plugin-terser": "1.0.0",
     "@rollup/plugin-typescript": "12.3.0",
-    "@tanstack/react-query": "5.90.21",
     "@types/node": "25.5.0",
     "@types/react": "19.2.14",
     "react": "19.2.4",
     "rollup": "4.59.0",
-    "rollup-plugin-sass": "1.15.3",
     "rollup-preserve-directives": "1.1.3",
     "tslib": "2.8.1",
     "typescript": "5.9.3"
   },
   "peerDependencies": {
     "@atview/core": ">= 0.0.1",
-    "@atview/ui": ">= 0.0.1",
-    "@atview/react": ">= 0.0.1",
-    "@atview/xrpc": ">= 0.0.1",
-    "@tanstack/react-query": ">= 5.0.0",
     "react": ">= 19.1.0"
   }
 }
