+++ packages/react/tsconfig.cjs.json
@@ -1,0 +1,10 @@
+{
+  "extends": "./tsconfig.json",
+  "compilerOptions": {
+    "module": "ESNext",
+    "moduleResolution": "bundler",
+    "outDir": "dist/cjs",
+    "declarationDir": "dist/cjs"
+  },
+  "include": ["src/**/*"]
+}
