+++ packages/react/tsconfig.json
@@ -1,0 +1,21 @@
+{
+  "compilerOptions": {
+    "target": "ESNext",
+    "module": "es2020",
+    "jsx": "react",
+    "esModuleInterop": true,
+    "forceConsistentCasingInFileNames": true,
+    "strict": true,
+    "skipLibCheck": true,
+    "outDir": "dist/cjs",
+    "rootDir": "src",
+    "declaration": true,
+    "declarationDir": "dist/cjs",
+    "moduleResolution": "bundler",
+    "paths": {
+      "@src/*": [
+        "./src/*"
+      ]
+    }
+  }
+}
