+++ packages/ui/src/features/shared/ui/blockquote/index.tsx
@@ -3,7 +3,7 @@ import React from "react";
 import "./blockquote.scss";
 
 export interface BlockquoteProps {
-    children: React.ReactNode;
+    children?: React.ReactNode;
 }
 
 export const Blockquote: React.FC<BlockquoteProps> = (props) => <blockquote className="atview-blockquote" {...props} />;
