+++ packages/ui/src/features/shared/ui/bold/index.tsx
@@ -3,7 +3,7 @@ import React from "react";
 import "./bold.scss";
 
 export interface BoldProps {
-    children: React.ReactNode;
+    children?: React.ReactNode;
 }
 
 export const Bold: React.FC<BoldProps> = (props) => <b className="atview-bold" {...props} />;
