+++ packages/ui/src/features/shared/ui/highlight/index.tsx
@@ -3,7 +3,7 @@ import React from "react";
 import "./highlight.scss";
 
 export interface HighlightProps {
-    children: React.ReactNode;
+    children?: React.ReactNode;
 }
 
 export const Highlight: React.FC<HighlightProps> = (props) => <mark className="atview-highlight" {...props} />;
