+++ packages/ui/src/features/shared/ui/inline-code/index.tsx
@@ -3,7 +3,7 @@ import React from "react";
 import "./inline-code.scss";
 
 export interface InlineCodeProps {
-    children: React.ReactNode;
+    children?: React.ReactNode;
 }
 
 export const InlineCode: React.FC<InlineCodeProps> = (props) => <code className="atview-inline-code" {...props} />;
