+++ packages/ui/src/features/shared/ui/italic/index.tsx
@@ -3,7 +3,7 @@ import React from "react";
 import "./italic.scss";
 
 export interface ItalicProps {
-    children: React.ReactNode;
+    children?: React.ReactNode;
 }
 
 export const Italic: React.FC<ItalicProps> = (props) => <i className="atview-italic" {...props} />;
