+++ packages/ui/src/features/shared/ui/paragraph/index.tsx
@@ -5,7 +5,7 @@ import { Typography } from "../typography";
 import "./paragraph.scss";
 
 export interface ParagraphProps {
-    children: React.ReactNode;
+    children?: React.ReactNode;
 }
 
 export const Paragraph: React.FC<ParagraphProps> = (props) => <Typography className="atview-paragraph" {...props} />;
