+++ packages/ui/src/features/shared/ui/strikethrough/index.tsx
@@ -3,7 +3,7 @@ import React from "react";
 import "./strikethrough.scss";
 
 export interface StrikethroughProps {
-    children: React.ReactNode;
+    children?: React.ReactNode;
 }
 
 export const Strikethrough: React.FC<StrikethroughProps> = (props) => <s className="atview-strikethrough" {...props} />;
