+++ packages/ui/src/features/shared/ui/table-cell/index.tsx
@@ -3,7 +3,7 @@ import React from "react";
 import "./table-cell.scss";
 
 export interface TableCellProps {
-    children: React.ReactNode;
+    children?: React.ReactNode;
     colSpan?: number;
     rowSpan?: number;
 }
