+++ packages/ui/src/features/shared/ui/table/index.tsx
@@ -3,7 +3,7 @@ import React from "react";
 import "./table.scss";
 
 export interface TableProps {
-    children: React.ReactNode;
+    children?: React.ReactNode;
 }
 
 export const Table: React.FC<TableProps> = (props) => (
