+++ packages/ui/src/features/shared/ui/task-list/index.tsx
@@ -3,7 +3,7 @@ import React from "react";
 import "./task-list.scss";
 
 export interface TaskListProps {
-    children: React.ReactNode;
+    children?: React.ReactNode;
 }
 
 export const TaskList: React.FC<TaskListProps> = ({ children }) => {
