+++ packages/ui/src/features/shared/ui/underline/index.tsx
@@ -3,7 +3,7 @@ import React from "react";
 import "./underline.scss";
 
 export interface UnderlineProps {
-    children: React.ReactNode;
+    children?: React.ReactNode;
 }
 
 export const Underline: React.FC<UnderlineProps> = (props) => <u className="atview-underline" {...props} />;
