+++ packages/ui/src/features/shared/ui/unordered-list/index.tsx
@@ -3,7 +3,7 @@ import React from "react";
 import "./unordered-list.scss";
 
 export interface UnorderedListProps {
-    children: React.ReactNode;
+    children?: React.ReactNode;
 }
 
 export const UnorderedList: React.FC<UnorderedListProps> = ({ children }) => {
